+++ apps/doji-frontend/src/modules/session-history/components/SessionHistoryCard/index.tsx
