+++ libs/mui/src/lib/index.ts
@@ -7,5 +7,4 @@ export * from './ReviewMenu'
 export * from './Logo'
 export * from './CompactProfile'
 export * from './SessionCard'
-export * from './SessionHistoryCard'
 export * from './Achievement'
